arreglo de bugs

+++ integrador2/src/main/java/dto/CarreraDTO.java
@@ -3,12 +3,12 @@ package dto;
 public class CarreraDTO {
     int idCarrera;
     String nombre_carrera;
-    int cantidad_inscriptos;
+    long cantidad_inscriptos;
 
     public CarreraDTO() {
     }
 
-    public CarreraDTO(int idCarrera, String nombre_carrera, int cantidad_inscriptos) {
+    public CarreraDTO(int idCarrera, String nombre_carrera, long cantidad_inscriptos) {
         this.idCarrera = idCarrera;
         this.nombre_carrera = nombre_carrera;
         this.cantidad_inscriptos = cantidad_inscriptos;
@@ -22,7 +22,7 @@ public class CarreraDTO {
         return nombre_carrera;
     }
 
-    public int getCantidad_inscriptos() {
+    public long getCantidad_inscriptos() {
         return cantidad_inscriptos;
     }
